Aggiunte classi per le entità (con parametri provvisori) e corretto uri di neo4j in Settings.xml

+++ src/main/java/it/unipi/iit/inginf/lsmdb/communitunes/entities/Entity.java
@@ -1,5 +1,7 @@
 package it.unipi.iit.inginf.lsmdb.communitunes.entities;
 
 abstract class Entity {
-    public int ID;
+
+    public long ID;
+
 }
